Refactor views around building tree view
The old page blocks view would return a breadth first view of the note
tree. In practice this was not going to work since nested blocks need to
be seen before siblings to a parent block. By adding a new positions
ancestry chain its possible to put the blocks into an order that will be
more proper for the UI to render.

There could be an issue in the key getting way too long if the positions
are allowed to get too long. This will get fixed when the logic that
handles reordering blocks makes sure their position stays within a
reasonable range.